+++ packages/adapter-begin/index.d.ts
@@ -1,1 +1,0 @@
-export default function (): import('@sveltejs/kit').Adapter;