+++ packages/kit/src/core/config/options.js
@@ -39,6 +39,18 @@ const options = object(
 				return input;
 			}),
 
+			alias: validate({}, (input, keypath) => {
+				if (typeof input !== 'object') {
+					throw new Error(`${keypath} should be an object`);
+				}
+
+				for (const key in input) {
+					assert_string(input[key], `${keypath}.${key}`);
+				}
+
+				return input;
+			}),
+
 			// TODO: remove this for the 1.0 release
 			amp: error(
 				(keypath) =>